Move the `AsyncGroup` into a directory
