+++ src/tribler/core/components/restapi/rest/rest_endpoint.py
@@ -6,7 +6,7 @@ from typing import Dict, TYPE_CHECKING
 
 from aiohttp import web
 
-from tribler.core.utilities.async_group import AsyncGroup
+from tribler.core.utilities.async_group.async_group import AsyncGroup
 
 if TYPE_CHECKING:
     from tribler.core.components.restapi.rest.events_endpoint import EventsEndpoint
